requirements.txt
gymnasium==0.29.1
python-chess==1.999
torch==2.9.0
numpy>=1.24
huggingface-hub>=0.24

